[FIX] pos_preparation_display: prevent keyerror when internal note is empty
When a customer adds an empty internal note and tries to order the product,
a traceback will appear.

Steps to reproduce the error:
- Open PoS restaurant session
- Add any product > Order
- Add same product again > Add an empty internal note
- Order

Error: A traceback appears:
"KeyError: (3, '', '[]')"

https://github.com/odoo/enterprise/blob/4077a5e39fba7197d32d5db05171ad9c8f98c392/pos_preparation_display/models/pos_order.py#L95
Here when customer adds an empty internal note, quantity_data will become empty,
So, it will lead to the above traceback.

sentry-5494183916

closes odoo/enterprise#66145

Related: odoo/odoo#172730
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>