[IMP] website_enterprise, *: Adapts appointment for new dragNDrop system
*: website_appointment

Following the change in the way of dragging and dropping a snippet onto
the page (by selecting it from a dialog displaying the actual previews
of snippets). The "website_appointment" app has been adapted to this
change in this commit.

Data has also been added in this commit so the dynamic appointment
snippet can be shown in the snippet dialog.

task-3919405

closes odoo/enterprise#66150

Related: odoo/odoo#166967
Related: odoo/design-themes#817
Signed-off-by: Robin Lejeune (role) <role@odoo.com>