[FIX] sign: item disappears out-of-bounds on drag
Before this commit, after we drag a sign item out of bounds, it would disappear
until we refresh the page. This was happening because we were hiding the item at
drag start, then it would only get re-rendered after dropping it inside the page.
Thus, if we dragged the item out of bounds instead of inside the page, the item
would never get re-rendered.

After this commit, the sign item can be dragged anywhere, even out-of-bounds. It
works properly because now when the item is dragged out-of-bounds, we make the item
visible again even though the the invalid operation is discarded.

task-4037888

closes odoo/enterprise#66155

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>