[PERF] mrp_workorder: fetch operation note when necessary in Shop Floor
Before this commit, the operation note was fetched and sanitized for each work orders.
This induces performance issues as users tend to have base64 image encoded in their operation note.
Moreover, it was not necessary as the operation note is only displayed
when the user clicks on the worksheet or the quality check.

This commit changes the behavior to fetch the operation note only when
the user clicks on the worksheet or the quality check.
It also improves the performance by fetching the operation note only once and caching it.

opw-3983324

closes odoo/enterprise#67748

X-original-commit: 991416cbfe5aff6bb339c1394190b42f17ec5b63
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Julien Segers (juse) <juse@odoo.com>