[IMP] account_reports: Annotalypse
Description of the issue/feature this commit addresses:

The current behavior of footnotes makes it impossible to have multiple
footnotes by company by line. Also, when annoting a line in a report for a
certain period, and changing the period of the data shown by the reports, the
annotation stays. This leads to footnotes being irrelevant or
give false pieces of information.

This commit transforms footnotes into annotations which are much more
flexible and are a more powerful tool for the users.

---

Desired behavior after the commit is merged :

This commit brings a rework to the annotations' system by moving adding
and changing some features they bring.

1 - The annotation now disposes of a "date" field which is set to the value
of the "date_to" period of the data currently shown on the report. This
value is then used during the computation of which annotations have to be
shown and which have not to be shown so that only annotations for
the current period appear.

2 - When clicking on the "Annotate" button on a line that already has an
annotation, previous behavior was the opening of a dialog with the content
of the current annotation which gave the possibility to edit it.
Now, if a line has already one visible annotation, that button is not shown
anymore. The line name has a little comment bubble next to it which opens the
"annotation popover". Inside the popover, the user can read mulitple pieces of
information regarding each annotation of the line which are : its text content,
its author, the date it has been written at and the date it targets. The user
can also edit and delete each of these annotation or add more via a "New"
button.

3 - The footnote section at the bottom of a report with at least one footnote
has been removed as the content of the annotation is now available directly on
the annotated line and it made no sense to leave it.

4 - In the xlsx export, an additional column has been added to
keep the annotations during the export. There is one column by line
regardless to the number of annotations the line has and in a case in
which a line would have multiple annotations, each of them would be
numbered and split by a return character.

5 - In the PDF export, the rendering is now able to show multiple
annotations on a single line.

6 - Journal Report and Followup Report used to have a footnote rendering
process for the pdf export but since those Reports don't offer the
possibility to annotate lines, thoses process have been removed.

Adding this improvement, annotations are much more flexible
and can bring real value to reports.

---

task-3301125

closes odoo/enterprise#48530

Related: odoo/upgrade#6244
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>