[FIX] iap_extract: fill document with the correct company context
When the document is filled in with the OCR values, we need to use the
document's company context, otherwise it can lead to inconsistent values
being selected.
For example, the payment terms of a supplier are company dependent, so
we need the right company context to select the one from the company of
the invoice.

Related task: #3902741

closes odoo/enterprise#66168

Signed-off-by: William André (wan) <wan@odoo.com>