[IMP] approvals_purchase: simplify 'purchase' approvals
When we make an approval, of 'purchase' type, and 'create rfq',
the products are directly purchased from the top vendor of the list.
If you want to change that, it will be time consuming to remove
it from an open RFQ, or change the vendor, find the right price.

instead, in this commit, we allow to indicate vendor straight away
from the product tab of the approvals.

task-3871938

closes odoo/enterprise#66254

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>