[FIX] l10n_us_payment_nacha: left-justify company data/identification
The Company Identification field was incorrectly formatted as a
numeric field. It is specified as numeric by some banks [1], but the
official specification is alphanumeric [2].

On top of that the alphanumeric Company Discretionary Data field was
right-justified. A minority of banks don't handle this correctly.

There's no official specification on how to align alphanumeric fields
[1]. The decision to left-justify is based on the following:

- Chase specifies to left-justify them [3]
- The most extensive open-source ACH library left-justifies them [4]
- Another Python ACH library left-justifies them too [5]
- Online images of ACH files seem to left-justify them as well

The only two remaining fields that are right-justified with
spaces ({:>...}) are Immediate Destination and Immediate Origin. This
is on purpose. Although they are numeric fields and thus right
aligned, they are specified to start with a blank space, not a leading
0 [2].

[1] https://web.archive.org/web/20230624090124/https://files.nc.gov/ncosc/documents/eCommerce/bank_of_america_nacha_file_specs.pdf
[2] https://achdevguide.nacha.org/ach-file-overview
[3] https://www.chase.com/content/dam/chaseonline/en/demos/cbo/pdfs/cbo_nacha_filespecs.pdf
[4] https://github.com/moov-io/ach/blob/073d011f811605e1b6051792163159a55feca533/converters.go#L81-L94
[5] https://github.com/travishathaway/python-ach/blob/cc8b6448d75f3815c91d3b8a5dadd6731aa96e71/ach/data_types.py#L57-L76

opw-4033830

closes odoo/enterprise#66451

X-original-commit: 13657ff494626697b170a02a2c17dd9415ed5e84
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>