[FIX] l10n_ph_reports: adapt xlsxwriter usage for ubuntu noble
The `col_sizes` attribute on an worksheet was removed in xlsxwriter
version 3.0.6 [0]. Ubuntu Noble provides xlsxwriter 3.1.9.
The xlsxwriter author advises [1] to use the `col_info` attribute
instead, but this attribute appeared in 3.0.6 too (rename of `colinfo`).

With this commit, the version is checked in order to use the right
attribute.

[0]: jmcnamara/XlsxWriter@860f4a2
[1]: jmcnamara/XlsxWriter#937

Related: odoo/odoo#171924
Related: odoo/odoo#173788
Part-of: odoo/enterprise#66974
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>