[FIX] mrp_mps: do not crop big numbers
When filling big numbers (10+ digits) in the input field with Actual Demand/Replenishment activated,
the field limits itself to 38% of the cell width, which then crops the number inside.

This fix removes the inline flex and the max-width limitation while adding a small margin-top.
This way, if Actual Demand/Replenishment is activated, it will be shown over 2 lines in the cell.

closes odoo/enterprise#66164

Signed-off-by: Steve Van Essche <svs@odoo.com>