[FIX] mrp_mps: do not crop big numbers
When filling big numbers (10+ digits) in the input field with Actual Demand/Replenishment activated,
the field limits itself to 38% of the cell width, which then crops the number inside.

This fix removes the inline flex and the max-width limitation while adding a small margin-top.
This way, if Actual Demand/Replenishment is activated, it will be shown over 2 lines in the cell.

closes odoo/enterprise#69069

X-original-commit: 2764e584c18d0db7c7ae653582f5c43af5342f35
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>