[FIX] stock_barcode: don't play sounds when params
The config parameter `stock_barcode.mute_sound_notifications`, who
allows to mute all sounds in the Barcode app, is not used on the
MainMenu, which means scanning a barcode on the main menu will always
play a sound regardless the configuration.
This commit fixes that.

Also, moves the code causing an RPC (`hasGroup`, `searchCount`) from the
JS to the python, so only one RPC is done (the call to
"/stock_barcode/get_main_menu_data" route.)

closes odoo/enterprise#66371

X-original-commit: 886d278c0728211d60e025492156a38827f07506
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>