[FIX] account_intrastat: comparison filter
Steps to reproduce
  - Display the intrastat report with at least one line
     - Create a european company
     - Install intrastat
     - Create and post an invoice to partner from another
       european country
     - Open the intrastat report

  - Activate the Comparison filter (with any other date)
  - Expand a line (if lines are grouped)
  => Traceback stating that the query has syntax error

Cause of the issue:
Missing parentheses between the different sections of the
UNION ALL statement.

This commit also fix issues related to the column_group_options
where duplicate lines existed when using the comparison filter.
It also cleans a grouped line (the section cell) if not data
should be displayed.

task-4017165

closes odoo/enterprise#67148

X-original-commit: 8d8ad1a81cf202104efc88a242c76737153ee24f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>