[FIX] stock_barcode: avoid duplicate sortIndex
In the Barcode app, lines objects have a `sortIndex` string value who is
used to sort them (lower are placed first, bigger are placed below.)
When a line is splitted, the new line get its sort index from the
original line but with + "1" to always be just after. The issue with
that is if the same line is splitted twice, both newly created lines
will have the same `sortIndex` which can lead to some unexpected result.

To avoid that, `sortIndex` is now a integer and we ensure that there
will be no duplicate by updating all lines `sortIndex` when needed.

closes odoo/enterprise#65182

Related: odoo/odoo#173521
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>