[REF] documents_spreadsheet: add navbar slot
For task 3761590, we also need to add an additional button
next to the spreadsheet name in the navbar, exactly like the Favorite
extension in `documents_spreadsheet`.

This commit adds an "official" extension slot. It's easier and less code
than creating a new component and xpathing.

Task: 3761590
Part-of: odoo/enterprise#65213
Related: odoo/odoo#170864
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>