[FIX] test_spreadsheet_edition: fix indeterministic test
The test `Selecting the cell with a unsolved thread does not open the
thread popover` was sometimes failing.

After investigation, it seems like the test is plain wrong. The thread
popover should open for unsolded threads, it's for resolved threads that
it should not open. Running the test in debug make it clear that the
threads popoved is opened. I'm not sure how the test was somehow
passing before.

closes odoo/enterprise#66296

Task: 4042515
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>