[FIX] sale_stock_renting: enforce rental transfer
**Current behavior:**
If a portal or public user causes a rental order to get
confirmed, rental transfer pickings won't get created if the DB
user has the setting enabled.

**Expected behavior:**
The transfer should be created with respect to the user of the
database, not the external users.

**Steps to reproduce:**
1. Enable rental transfers

2. Create a rental order to a portal user

3. In a fresh browser, open the preview link as the portal user

4. Sign & Pay for the order

5. See that no transfers got created

**Cause of the issue:**
We use the `env.user` to determine whether the transfers should
be generated.

**Fix:**
Use the creator of the order to decide to create the transfer.

opw-3933511

closes odoo/enterprise#66302

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>