[FIX] account_reports: Balance sheet test savepoint fix
Instead of calling `env.cr.savepoint()` in a loop, which will create a
savepoint at every iteration, we should create a single savepoint and
rollback to it at the end of every subtest.

In Postgres, the number of sub-transactions that may exist in active
transactions has a hard limit of 64 before sub-transactions are stored
on disk. It's therefore best practice to limit the number of savepoints
we create.

See also https://github.com/odoo/odoo/pull/167236

closes odoo/enterprise#66665

X-original-commit: f824dd17f864238f4aa88c7aa54532ec77a943a7
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>