[IMP] web_enterprise,*: Breadcrumbs in Navbar on mobile
*: spreadsheet_edition

Ususally, the Breadcrumbs representing the different steps of the user's
navigation is displayed in the ControlPanel. But, even while simplified
for smaller screens, this layout takes a lot of screen realestate on
smartphone-like screen sizes.

To reclaim this space in the ControlPanel and avoid as much as possible
the ControlPanel to wrap on multiple lines, this commit applies the
required adaptation in enterprise.

task-3336242

closes odoo/enterprise#65649

Related: odoo/odoo#171083
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>