[IMP] mrp_mps: open MPS from bom and see all children
This commit has 2 parts:

The BOM smart button to go to the MPS is now permanently visible and doesn't
show MPS records count. The count was removed for performance reasons, and
since the visibility was based on that count, we made it permanently visible.

Opening the MPS via that same button will show all child records and not just
the direct children. This is so that multi-level boms are correctly shown
for all levels and not just the first one.

task-4033339

closes odoo/enterprise#65989

Related: odoo/upgrade#6241
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>