[FIX] documents: do not use versioning if writing the same attachment
Steps to reproduce:
-------------------

- Install 'Documents' and 'Accounting' modules
- Go to document settings > Activate Accounting + go to "Journal"
  smart link then create a synchronized with Vendor Bills journal
- Go to accounting setting > Activate "Document Digitization" and
  select "Digitize automatically" for bills
- Send a PDF to the Vendor bill mail alias
- Open the created bill

Issue:
------

Error in the chatter.

Cause:
------

We try to set on the document an attachment already linked to it,
and because it goes through the versioning code, it will first link the
new attachment to the record (who was already the case), and then link
the old attachment (who is the same attachment as the new) to the
document.

Solution:
---------

If trying to set the same attachment already linked to the document,
we skip the versioning code.

opw-4034699

closes odoo/enterprise#66398

X-original-commit: aafd51fe41f9dd1b942d09c0e972916264b3bed2
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>