[FIX] planning: generate long monthly recurring slots
Steps
-----
[Planning]
- Create a shift from 1 July 2024 -> 31 July 2024 and chose the option
to repeat it every month. Save it.

1st issue: no shift is created in September.
2nd issue: November shift ends in December and overlaps with December
shift.

Cause
-----
- The 1st issue is caused by the first day of the September shift being
on a weekend. In case the first day is not a working day, a shift will
only be generated if the first day of the original shift was outside working
days.
https://github.com/odoo/enterprise/blame/adbf9375a07980c8b9b4f621dfaba5ffc99c4162/planning/models/planning_recurrency.py#L104-L107
- The 2nd issue happens because the length of all the repeated shifts is the
same than the original shift. In the case of a longer original month, this
will create overlapping generated shifts when we get to shorter months.

Introduced by 079b0ba241b8b4f7fbd74775e5fc43b27ca3c8f

Fix
-----
- Consider a shift with any day (and not just the starting one) outside of
working days as being outside working days.
- Adapt the length of the generated shifts in case of a monthly recurrence.
Take the minimum between the original shift duration in days and the number
of days in the month.

opw-3996123

closes odoo/enterprise#68002

X-original-commit: be09567a53e0399624e92fc98623843899c7c99b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>