[FIX] {documents_,}account{,_accountant}: adapt tests for audit trail
Since the audit trail is mandatory in Germany, we cannot remove moves or
any documents related.
We only need to remove the moves linked to the company of the tour.

We also have more queries because we now need to check if the move is
part of the audit trail before unlinking.

closes odoo/enterprise#66430

X-original-commit: ab81e11d325e246422decd1bf0a2d5c989074c7c
Related: odoo/odoo#172660
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>