[FIX] industry_fsm_sale: Correctly autofill salesperson on SO
Steps to reproduce:
- Projects > Field Service > New
- Pick a customer but no assignee
- Products > Add any product
- Back to task > Sale order > Other info tab

Expected behavior: salesperson is the first assignee on
the task if any, current user if not (first assignee as in the first
listed on the task, which sorts them alphabetically).

What happens:
No salesperson is set. The auto assigned salesperson is overwriten to be
the 1st assignee after SO creation, even when there is no assignee to
the task. Changed that behavior to only overwrite if there is something
to overwrite with.

opw-4015184
opw-4026074 (duplicate)

closes odoo/enterprise#66404

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>