[FIX] mrp_workorder: fix traceback on WO list view
To reproduce:
- Open manufacturing -> Operations -> Work Orders

Root cause:
Since the introduction of the Shop Floor in [1], a JS class was deleted
from the workorders list view, but we forgot to remove the attribute on
the view itself.
Since the changes in [2], this will now result in a hard crash when
trying to load the view (as there no longer is a null check).

This commit fixes this by removing the JS class attribute from the view.

[1] https://github.com/odoo/enterprise/pull/43610
[2] https://github.com/odoo/odoo/pull/171073

closes odoo/enterprise#66409

Signed-off-by: Steve Van Essche <svs@odoo.com>