[IMP] hr_payroll: Remove hr.salary.attachment.type model
Purpose
=======

Salary attachment types are nearly the same than other input type.

Actually for each salary attachment type, a other input type with the same code
should be defined as well.

The model itself can be removed and the additional information can be moved either
on the attachment, or on the other input type.

Taskid: 3988630
Part-of: odoo/enterprise#66413
Related: odoo/upgrade#6246
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>