[FIX] hr_appraisal : fix employee's deletion
Before this commit, you can't delete an employee linked to an appraisal, a goal, or goal's manager.
When you tried to do it, you had a traceback.
With this commit; when you want to delete an employee; his appraisals and his goals will be deleted too.
But if he was a goal's manager; he can't be delete he needs to be archived to keep an history.

closes odoo/enterprise#58076

Related: odoo/odoo#156521
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>