[IMP] iot: add is_websocket_active field
Before this PR, there was no easy way to identify if an IoT box was
connected with a websocket, potentially making debugging report printing
difficult.

This PR adds an `is_websocket_active` field to the IoT box model, which
defaults to False but is set to True when it subscribes to the
websocket. If the websocket is closed by Odoo, it is set to False again.

Task ID: 4045291

closes odoo/enterprise#66722

Related: odoo/odoo#173286
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>