[FIX] account_online_synchronization: Fix filtering of transactions with empty identifier
The aim of this commit is fixing the filtering of transactions
when transactions have an empty identifier.
Even if it's shouldn't be the case, it happens that providers
return transactions without an identifier. In this case, it will
always importe the first transaction and not the others as they have
a similar id (an empty string). Now, we ignore in the filter these
transactions.

opw-4014171

closes odoo/enterprise#66737

X-original-commit: 8188fa378ea994e9518a7f02e20556aa50d6587a
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>