[FIX] account_online_synchronization: Fix filtering of transactions with empty identifier
The aim of this commit is fixing the filtering of transactions
when transactions have an empty identifier.
Even if it's shouldn't be the case, it happens that providers
return transactions without an identifier. In this case, it will
always importe the first transaction and not the others as they have
a similar id (an empty string). Now, we ignore in the filter these
transactions.

opw-4014171

closes odoo/enterprise#66473

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>