[FIX] pos_*_preparation_display: split internal note when adding product
If you already ordered product A and add another product A with an
internal note. The internal note would end up on both product A. The
correct behavior would be to have the internal note on only one of them.

Steps to reproduce:
-------------------
* Setup a PoS restaurant and a preparation display
* Open PoS session
* Add 1 product A to the order and send it to preparation
* Add 1 product A, and add an internal note on it
* Send it to preparation
> Observation: On the preparation display both product will have the
internal note

Why the fix:
------------
To fix the issue, when adding an internal note we only add it on the
product that hasn't been sent in preparation and split the order line in
2 if necessary. We also take the order line uuid in account when
processing the orders in the preparation display to differentiate the
line correctly.

The behavior should be the following:
- If you add a note on a line that has no quantity sent in preparation
  the note will be on all the quantity
- If you add a note on a line that has all quantity sent in preparation
  the note will be on all the quantity
- If you add a note on a line that has a part of the quantity sent in
  preparation the note will be on the quantity not sent, and the line
  will be split in 2. One line with already sent, and one with the rest.

opw-4000386

closes odoo/enterprise#66477

Related: odoo/odoo#171913
Signed-off-by: David Monnom (moda) <moda@odoo.com>