[IMP] account_invoice_extract: set imported flag
When changing the product on the vendor bill line, the price and taxes are recomputed.
The corresponding community PR uses a flag to disable the computation for lines that have been imported.
Since the OCR feature writes the lines later, the flag is set explicitly on created lines.

Task-3837975

closes odoo/enterprise#66506

Related: odoo/odoo#172787
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>