[FIX] approvals_purchase: don't propagate default_name context key
Issue
-----
[Approvals, Purchase, Stock]
1. In approvals, set the current user as the approver for the "Create RFQ's"
approval type.
2. Approvals > New Request > Create RFQ's > Select a product with a vendor
setup ("Acoustic bloc screen" for example) > Submit > Approve > Create RFQ
> Click smart button to navigate to the PO > Confirm the PO (should work)
3. Repeat step 2, ValidationError when trying to confirm the PO.

Note: the issue only arises in this specific flow, the PO can still be confirmed
by for example just refreshing the page.

Issue
-----
A 'default_name' key is added to the context when creating a request for a
category which uses sequential names (which is the case for RFQ's approvals).
https://github.com/odoo/enterprise/blob/65a8e25f1ff2d0faf17ad73c1402615d320ffc48/approvals/models/approval_category.py#L154

As a result, it gets propagated all the way to the creation of the picking
when the PO is confirmed.
https://github.com/odoo/odoo/blob/beb8c7d86ae6062dbad9fcb17bc3141d14a9fcdc/addons/stock/models/stock_picking.py#L721
As pickings require a unique (name, company_id), a first picking can be created
with the name 'New', but later ones will fail.

Fix
-----
Avoid propagating the default_name key when opening the purchase orders so
the picking created when confirming the order will have the correct name.

opw-3990816

closes odoo/enterprise#68301

X-original-commit: b04db8e565dbdb443d827f979843f73002b993fb
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>