[FIX] helpdesk: allow quick_create from ticket kanban view
Currently, even tough the quick_create view are defined in the code
base, and defines on the kanban view, the quick create view is not used.

Step to reproduce:
- Install helpdesk
- Open helpdesk app
- Open the 'all tickets' or the 'My tickets' view
- Select the kanban view
- Click on the 'new' button

A ticket view form is opened.
Expected behavior : the quick create view form should be open in the
kanban view.

Source of the issue:
the 'on_create' attribute is missing in the kanban definition

task - 4047630
version 15.0 - master

closes odoo/enterprise#70513

X-original-commit: 36a8b8a2cd9f53651d05835120d3265d44bcdf21
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>