[FIX] helpdesk: allow quick_create from ticket kanban view
Currently, even tough the quick_create view are defined in the code
base, and defines on the kanban view, the quick create view is not used.

Step to reproduce:
- Install helpdesk
- Open helpdesk app
- Open the 'all tickets' or the 'My tickets' view
- Select the kanban view
- Click on the 'new' button

A ticket view form is opened.
Expected behavior : the quick create view form should be open in the
kanban view.

Source of the issue:
the 'on_create' attribute is missing in the kanban definition

task - 4047630
version 15.0 - master

closes odoo/enterprise#71220

X-original-commit: d3950de88546ea30f71563d2d4cb918c2a6d65c9
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>