[FIX] account_report: new condition for export report in file mode
When we export a report as a file, we don't want to manage the different column groups. For example, if comparisons are activated in the report, we only want to add the original values to the file, without any comparisons, as they don't make any sense in a file format other than PDF.

This commit ensures that if comparisons are activated when exporting a file, the options aren't re-computed, so comparison data is ignored during the export.

Part-of: odoo/enterprise#60153
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>