[REF] mail_enterprise, *: use real python model name in Store insert
Not having the correct name based on python model name makes it much
harder to implement future improvement such as auto-formatting of
relations due to python having no knowledge of JS model definitions.

This also allows to remove a lot of unnecessary type and model keys for
channel and persona, because they can be deduced from the model name.

Part of task-3605717

closes odoo/enterprise#66478

Related: odoo/odoo#172731
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>