[IMP] l10n_mx_edi_extended: city domain
When you select a city from the city list (city_id field) on the res.partner
the cities are filtered by country in standard. However this is complicated to
use because more often than not there are many cities with the same name. This
is specially frequent in large countries.
So simply changing it to domain="[('state_id', '=', state_id)]" solves the issue.

closes odoo/enterprise#66931

Task: 4027290
X-original-commit: be57478a763e81c1110ca3ec3ef8613b21d6ade0
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>