[FIX] l10n_cl_edi: support listaEventosDoc being a list
When parsing a claim status from the SII, the list of events in the
response can either be a single event or a list of events. If a list was
returned this would fail with a parsing error because the original
implementation only expects a single event. That then led to the status
getting repeatedly checked and logging "list indices must be integers or
slices, not str" when done through a cron.

This fix supports both the single and multiple event scenario. The
l10n_cl_claim field is not displayed anywhere by default and is only
used to check if the claim status still needs to be logged in the
chatter or not. Since multiple response codes are possible, in theory
it should be a list instead of a single value, but that would be more
appropriate for a change in master. A boolean flag would also suffice if
the response code doesn't need to be displayed separately. We currently
just set the last code that was returned.

The responses observed in the customer database include the code NCA,
which isn't expected by Odoo. This code was added as a valid one.

opw-3920273

closes odoo/enterprise#70184

X-original-commit: 039c1354e08416049a61423eb19ab8b964e22701
Signed-off-by: Merel Geens <mege@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>