[FIX] documents_account: skip versioning when switching main attachment
Steps to reproduce:
-------------------

- Install `documents_account` module
- Go to settings -> activate 'Centralize accounting files and documents'
- Still in settings, under centralization feature, click on `Journals`
  to create a journal to synchronize
- Create a record with `Vendor Bills` as journal and with any workspace
- Go to account module and create a vendor bill in draft
- Attach 2 files at once on the vendor bill
- Switch the main attachment on the vendor bill
- Go to list view and open the newly created vendor bill

Issue:
------

One attachment has been removed (not linked anymore to the move).

Cause:
------

When switching the main attachment, the versioning is triggered; the
new main attachment is set on the document and the previous is added
to the history for versioning (and therefore also change the res fields
to be linked to the document instead of the move).

Solution:
---------

When updating the attachment of a document that already have an
attachment (normal versioning), we should remove the link between the
current document attachment (not the new one) and the related model
only if the related model has not a `message_main_attachment_id` field
or if the current document attachment is different of the main
attachment set on the related model.

opw-4028789

X-original-commit: 1065f32
Part-of: odoo/enterprise#67464
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>