[FIX] documents: information of message sender from alias
Before this commit:
The information of message sender using email alias was not shown,
instead only mail subject with name Odoobot was displayed in chatter.

After this commit:
The information of sender such as email address, name, subject of email
and email body is not displayed in chatter along with the attachment.

Task - 4008634

closes odoo/enterprise#66503

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>