[FIX] web_studio: prevent demo data used inside reports outside studio
Steps to reproduce
==================

- Install hr_skills,web_studio
- Go to Employees
- Open any record
- Add a new resume entry without description
- Save
- Click on the actions gear
- Print > Print Resume

=> `Odoo India pvt. Ltd` is displayed in the report in place of the
   missing description

Cause of the issue
==================

With the new report editor in 16.4, some demo data have been added
directly inside the report templates as default content.
The assumption was that it would be displayed inside the report editor,
but outside, since the field would have a value, the value would take
over.

The default content if displayed when the field value is falsy.

In this case, the default content is used instead of the empty
description.

Solution
========

Introduce a new attribute, data-oe-demo, to be used inside the report
editor. This doesn't have the drawbacks of the default content as we can
choose to only display it inside the report editor

---

[commit]: https://github.com/odoo/odoo/commit/a45c6c545ffdcee3bb50b3455884b8296e5d6a7c

opw-4033434

closes odoo/enterprise#66708

X-original-commit: 559aa39162c76b0c42464a327d220964e4a16fb0
Related: odoo/odoo#173263
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>