[FIX] pos_settle_due: get updated total_due
FIX n°1 Before this commit, the total_due was only updated upon loading
the session and when settling the customer's dues. This meant that if
the total_due was outdated and equal to 0, the "settle dues" button
would be invisible and we had no way of settling dues.

This commit ensures we also update the total_due when clicking on
"details" from the partner list screen, allowing us to see the "settle
dues" button and proceed with the settlement for any partner with dues.

Fix n°2 Before this commit, the get_total_due method did not take into
account the settled dues in an ongoing POS session. This meant that
despite "refreshing" the total_due after settlement, it wouldn't change.

This commit modifies get_total_due to look into ongoing POS sessions for
paid but unposted settled dues.

opw-3976254

closes odoo/enterprise#66369

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>