[IMP] whatsapp_pos: preserve input data on receipt_screen
Before this commit:
====================
When switching the send options (email, WhatsApp, SMS) on the receipt screen,
any manually entered input (such as email or phone number) would be erased if a
customer was not selected or if customer data (email or mobile number) was not
present.

Example:
Enter an email address manually.
Switch the send option to WhatsApp or SMS.
The manually entered email address would be erased.

After this commit:
=================
Manually entered input will be preserved when switching between send options
on the receipt screen.

Task-4038000

closes odoo/enterprise#66554

Related: odoo/odoo#172890
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>