[IMP] l10n_nl_reports: Adjustable XAF docRef len
Before this commit:
Some softwares do not follow the specification.
Our users cannot import the generated XAF file in other softwares because the docRef length is more than 35 characters.

After this commit:
We allow our users to create a system parameter to change the docRef length if they need to import the XAF file in another software.

opw-3792115
opw-3829459

closes odoo/enterprise#68151

X-original-commit: 41e655f4b5a8c827c83e5b59f2c71f8e9e8b8f76
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Corentin Thaon (thco) <thco@odoo.com>