[FIX] mrp_workorder: Display MO items to produce in full
Steps to reproduce:
- Manufacturing > Products > BoM
- Create BoM with at least 1 operation
- New MO with Quantity to produce >= 1000
- Shop Floor > Remove filters

What happens:
Work order card truncates number of units to produce by 3 digits, i.e.
a quantity of 2500 displays as 2, 34000 as 34, etc... This is caused by
the successive conversions in formatFloat, they go:
Number (quantityToProduce) > String (formatFloat) >
    Number (ParseFloat) > String (String())
Which looks to be an iffy solution to forcefully round the float into
an integer value.

What this fix does:
Getting rid of the superfluous conversions fixes the truncating issue.

A precision was passed all along and simply failed to display because of
the aformentioned issue. So now 2500 displays as 2500.00 this is
probably preferable as it was explicitly passed as props to he display.

The default behavior which used to show integers is now changed to 2 digits
precision (as defined by formatFloat) so use cases where that is not
desirable will need to pass their own precision as props.

Another instance of this conversion stack was also modified as it is
most likely subject to the same issue.

opw-4040457

closes odoo/enterprise#66570

Signed-off-by: William Henrotin (whe) <whe@odoo.com>