[FIX] account_accountant: pass missing parameters to tax calculation
Invoice creation through reconciliation models was implemented
in #59555, but the call to function
_eval_taxes_computation_prepare_context had missing paramenters.

When calculating the price unit for invoice creation, the function
_get_invoice_price_unit_from_price_total prepares the context
for tax calculation with _eval_taxes_computation_prepare_context.
In the latter function call, the parameters rounding_method and
precision_rounding were not being passed. This commit fixes that.

closes odoo/enterprise#66557

Signed-off-by: Laurent Smet (las) <las@odoo.com>