[FIX] test_spreadsheet_edition: fix comment popover test
This test wants to show the behavior of hovering a cell with
a comment thread.
But the test creates a comment thread on A1 (which is selected by default).

Selecting a cell with a comment thread also opens the popover, which
interferes with the behavior we want to test.

This commit moves the thread and hovering to A2, which is not selected.

closes odoo/enterprise#66584

X-original-commit: 152dd19be3390c9ed23f53316fa06b50ee2708dd
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>