[FIX] account_reports: Annotations on tax / cash flow report
The aim of this commit is fixing the traceback when user try to
annotate a line in the tax or cash flow report.
This traceback was due by the fact that we basically reimplement
the _build_line_id not correctly to remove a part of the id.
Now, the code is using the existing _build_line_id.
In the same time, this commit implement the buildLineId and parseLineId
functions in javascript.

task-4045835

X-original-commit: 0733c7412fcdc64b7f2c88ef1a5e2bb0866a353c
Part-of: odoo/enterprise#67782
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>