[FIX] iot: use one websocket channel instead of one per company
Using distinct websockets (ws) per company cause a lot of diffent
issues, for example:

Assume we have an IoT with no company set. In this case, it should
be able to print from any company.
In practice, it is not the case as the IoT will listen to the
channel of the first company, so it will always fail if the user
company is not the first one.

Same goes if we have child companies, if C2 is a child company of
C1, then C2 should be able to print too. With current code it is
not the case as C1 and C2 will have distincts channels

When the IoT start, it makes a call to know its channel
(see iot/setup route).
However, if we change the company set on the IoT, it should change
its channel to match the one of newly set company.
In practice it is not the case and will only happen on the next
iot/setup call (which happen each 3 seconds call in theory IF new
devices are detected)

Help simplifying the code as it was sometime complicated to be get
the right websocket.

For example, when the IoT reply back that the operation completed,
see `/iot/printer/status` route. Relying on the IoT device company
might be wrong as if no company is set, we don't know the company
that the user who did the request is, so we might send on the bus
on the wrong channel.
In this case, the document did print but the pop-up will still pop

The iot.channel had no practical way to be browsed as there is no
view defined for this model. The only feasible way to get its data
for a functional would be to run a server action or create a view.

Using a system parameter is more practical and only accessible for
admin users

- system parameter could be modified manually
- websocket send/receive more messages
- stability: to ease the transition, I tried to get the existing
ws channel using basic heuristics, but it might be wrong (if a
company other than the first is set on the IoT).
In the worst case, they will have to restart their IoT to get the
new channel
- refactor/clean of the code in master

opw-4038176

closes odoo/enterprise#66573

Signed-off-by: Loan Sens (lse) <lse@odoo.com>