[FIX] account_reports: Create/Save properly when popover click away
The aim of this commit is fixing a bug that occurs when user
leaves the popover during the creation of a new annotation.
It was caused by the fact that we waited for the request made via
orm.call even if the component is destroyed. It couldn't work.

Now, we're using the service directly from the environment (which
avoid to depend of the component status) and we're waiting for the
end of the promise before destroying definitely the component.
By doing that, the user can see the newly created annotation appears.
Before, it was not the case (even if the annotation was actually created).

task-4045835

X-original-commit: a76947ac876611b9b554e0d3a69c7dba50f7327d
Part-of: odoo/enterprise#67782
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>